S24 Release
Update for S24 version.
Example for the Node API can be found within plugins\example.nodes folder.

+++ plugins/cinema4dsdk/source/object/objectdata_latticeplanemodifier.cpp
@@ -72,7 +72,7 @@ namespace LatticePlaneModifierHelper
 		Float res = 1.0f;
 
 		for (Int32 i = 1; i <= k; i++)
-			res *= (n - (k - i)) / (float)i;
+			res *= Float((n - (k - i)) / (Float)i);
 
 		return res;
 	}
